requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
python-multipart==0.0.6
pandas==2.2.2
joblib==1.3.2
numpy==1.26.4
python-dotenv==1.0.0
redis==5.0.0
rq==1.14.0
slowapi==0.1.4
pyjwt==2.8.0
requests==2.31.0
sqlmap==1.6.14
